image: create from Picture methods
Move the Image constructors to methods of Picture<PictureSync>, since it
is the only source for images.

Split into two methods for construction instead of relying on a boolean
to try and derive the image. The problem with the boolean was that there
was no way to know whether the derive operation succeeded or not.