surface: add initial support for memory descriptors
We will eventually want to create surfaces with memory backed by DMABUFs
or other memory types. Introduce the SurfaceMemoryDescriptor trait that
can add the attributes required to assign memory to a VA surface, and
change the Surface constructor to take a slice of these descriptors
instead of the number of surfaces to create.

Regular VA buffers can be created by using () as the memory descriptor
type.